Putting tail_fitter into refreq, further cleaning

+++ triqs/gfs/gf/gf.hxx
@@ -135,7 +135,7 @@ namespace triqs {
       using zero_t            = zero_regular_t;
 
       // FIXME : std::array with NDA
-      using target_shape_t = arrays::mini_vector<long, Target::rank>;
+      using target_shape_t = arrays::mini_vector<int, Target::rank>;
 
       struct target_and_shape_t {
         target_shape_t _shape;
@@ -760,7 +760,7 @@ namespace triqs {
       using zero_t            = zero_view_t;
 
       // FIXME : std::array with NDA
-      using target_shape_t = arrays::mini_vector<long, Target::rank>;
+      using target_shape_t = arrays::mini_vector<int, Target::rank>;
 
       struct target_and_shape_t {
         target_shape_t _shape;
@@ -1357,7 +1357,7 @@ namespace triqs {
       using zero_t            = zero_const_view_t;
 
       // FIXME : std::array with NDA
-      using target_shape_t = arrays::mini_vector<long, Target::rank>;
+      using target_shape_t = arrays::mini_vector<int, Target::rank>;
 
       struct target_and_shape_t {
         target_shape_t _shape;
